// pro/src/apiClient/v1/models/LocationResponseModelV2.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LocationResponseModelV2 = {
  banId: (string | null);
  city: string;
  departmentCode: (string | null);
  id: number;
  inseeCode: (string | null);
  isManualEdition: boolean;
  isVenueLocation: boolean;
  label: (string | null);
  latitude: number;
  longitude: number;
  postalCode: string;
  street: (string | null);
};